refactor(server): route textGeneration OmpRpcRuntime/OmpSpawnError through the barrel
